Revert "feat: cpu player (#18)"
This reverts commit 56dce98ae23650adf87b25f28963e70cfb108771.
