+++ package.json
@@ -4,7 +4,7 @@
     "type": "git",
     "url": "https://github.com/sahilsinghrana/tictactoe.js.git"
   },
-  "version": "1.0.0-beta",
+  "version": "0.1.1",
   "description": "Tic Tac toe",
   "homepage": "https://github.com/sahilsinghrana/tictactoe.js",
   "main": "./dist/index.js",
@@ -20,8 +20,7 @@
   "scripts": {
     "doc": "typedoc",
     "dev": "tsup --watch",
-    "build": "tsup",
-    "test": "vitest"
+    "build": "tsup"
   },
   "files": [
     "/dist"
@@ -42,7 +41,6 @@
     "terser": "^5.30.2",
     "tsup": "^8.0.2",
     "typedoc": "^0.25.12",
-    "typescript": "^5.4.3",
-    "vitest": "^1.6.0"
+    "typescript": "^5.4.3"
   }
 }
