+++ src/index.ts
@@ -1,4 +1,2 @@
 import { TicTacToe } from "./lib";
 export default TicTacToe;
-
-console.log("I assssm linkedaawwsss");
